README symlink swap for simpler dist
README content in README file and README.md is just symlink for opening in smart IDEs and GitHub to determine markdown syntax. Now there is no need to include symlink in dist. Thanks @hroncok for suggestion. Another option would be `.github` file with `[README]` block, but IDEs are not so smart and symlink is more readable.